+++ system/src/Grav/Common/Markdown/Parsedown.php
@@ -10,6 +10,7 @@
 namespace Grav\Common\Markdown;
 
 use Grav\Common\Page\Interfaces\PageInterface;
+use Grav\Common\Page\Markdown\Excerpts;
 
 class Parsedown extends \Parsedown
 {
@@ -18,12 +19,21 @@ class Parsedown extends \Parsedown
     /**
      * Parsedown constructor.
      *
-     * @param PageInterface $page
+     * @param Excerpts|null $excerpts
      * @param array|null $defaults
      */
-    public function __construct($page, $defaults)
+    public function __construct($excerpts = null, $defaults = null)
     {
-        $this->init($page, $defaults);
+        if (!$excerpts || $excerpts instanceof PageInterface || null !== $defaults) {
+            // Deprecated in Grav 1.6.10
+            if ($defaults) {
+                $defaults = ['markdown' => $defaults];
+            }
+            $excerpts = new Excerpts($excerpts, $defaults);
+            user_error(__CLASS__ . '::' . __FUNCTION__ . '($page, $defaults) is deprecated since Grav 1.6.10, use new ' . __CLASS__ . '(new ' . Excerpts::class . '($page, [\'markdown\' => $defaults])) instead.', E_USER_DEPRECATED);
+        }
+
+        $this->init($excerpts, $defaults);
     }
 
 }