+++ system/src/Grav/Common/Markdown/ParsedownExtra.php
@@ -10,6 +10,7 @@
 namespace Grav\Common\Markdown;
 
 use Grav\Common\Page\Interfaces\PageInterface;
+use Grav\Common\Page\Markdown\Excerpts;
 
 class ParsedownExtra extends \ParsedownExtra
 {
@@ -18,14 +19,23 @@ class ParsedownExtra extends \ParsedownExtra
     /**
      * ParsedownExtra constructor.
      *
-     * @param PageInterface $page
+     * @param Excerpts|null $excerpts
      * @param array|null $defaults
      * @throws \Exception
      */
-    public function __construct($page, $defaults)
+    public function __construct($excerpts = null, $defaults = null)
     {
+        if (!$excerpts || $excerpts instanceof PageInterface || null !== $defaults) {
+            // Deprecated in Grav 1.6.10
+            if ($defaults) {
+                $defaults = ['markdown' => $defaults];
+            }
+            $excerpts = new Excerpts($excerpts, $defaults);
+            user_error(__CLASS__ . '::' . __FUNCTION__ . '($page, $defaults) is deprecated since Grav 1.6.10, use new ' . __CLASS__ . '(new ' . Excerpts::class . '($page, [\'markdown\' => $defaults])) instead.', E_USER_DEPRECATED);
+        }
+
         parent::__construct();
 
-        $this->init($page, $defaults);
+        $this->init($excerpts, $defaults);
     }
 }