+++ system/src/Grav/Common/Plugins.php
@@ -133,12 +133,25 @@ class Plugins extends Iterator
      */
     public static function all()
     {
-        $plugins = Grav::instance()['plugins'];
+        $grav = Grav::instance();
+        $plugins = $grav['plugins'];
         $list = [];
 
         foreach ($plugins as $instance) {
             $name = $instance->name;
-            $result = self::get($name);
+
+            try {
+                $result = self::get($name);
+            } catch (\Exception $e) {
+                $exception = new \RuntimeException(sprintf('Plugin %s: %s', $name, $e->getMessage()), $e->getCode(), $e);
+
+                /** @var Debugger $debugger */
+                $debugger = $grav['debugger'];
+                $debugger->addMessage("Plugin {$name} cannot be loaded, please check Exceptions tab", 'error');
+                $debugger->addException($exception);
+
+                continue;
+            }
 
             if ($result) {
                 $list[$name] = $result;
@@ -185,24 +198,31 @@ class Plugins extends Iterator
         $grav = Grav::instance();
         $locator = $grav['locator'];
 
-        $filePath = $locator->findResource('plugins://' . $name . DS . $name . PLUGIN_EXT);
-        if (!is_file($filePath)) {
+        $file = $locator->findResource('plugins://' . $name . DS . $name . PLUGIN_EXT);
+
+        if (is_file($file)) {
+            // Local variables available in the file: $grav, $config, $name, $file
+            $class = include_once $file;
+
+            $pluginClassFormat = [
+                'Grav\\Plugin\\' . ucfirst($name). 'Plugin',
+                'Grav\\Plugin\\' . Inflector::camelize($name) . 'Plugin'
+            ];
+
+            foreach ($pluginClassFormat as $pluginClass) {
+                if (class_exists($pluginClass)) {
+                    $class = new $pluginClass($name, $grav);
+                    break;
+                }
+            }
+        } else {
             $grav['log']->addWarning(
                 sprintf("Plugin '%s' enabled but not found! Try clearing cache with `bin/grav clear-cache`", $name)
             );
             return null;
         }
 
-        require_once $filePath;
-
-        $pluginClassName = 'Grav\\Plugin\\' . ucfirst($name) . 'Plugin';
-        if (!class_exists($pluginClassName)) {
-            $pluginClassName = 'Grav\\Plugin\\' . $grav['inflector']->camelize($name) . 'Plugin';
-            if (!class_exists($pluginClassName)) {
-                throw new \RuntimeException(sprintf("Plugin '%s' class not found! Try reinstalling this plugin.", $name));
-            }
-        }
-        return new $pluginClassName($name, $grav);
+        return $class;
     }
 
 }